Add files via upload
This repository contains a Jupyter Notebook analyzing crime rates in the United States using the 2009 UC Irvine ‘Communities and Crime’ dataset. Multiple linear regression (MLR) techniques are applied in order to identify the most significant socio-economic and demographic predictors of crime rates. The goal is to develop a predictive model focusing on the key factors influencing crime rates across the United States’ communities.